Convert all bytes to ki- variants
Transmission only provides units for Kibibytes and larger. This means
that when using their units, you'd have really strange things going on.

Simplest solution is to start from kilobytes. This is changing the way
the app runs for one specific client, but when µTorrent comes later, I'm
happy to have the smallest size be KiB. It's very rare you have files
where the sizes are measured in bytes, anyway.